+++ changelog.md
@@ -3,6 +3,9 @@ If you can't access the Google doc, here is a [PDF](https://github.com/crnormand
 
 This is what we are currently working on:
 
+- Fixed /status command when token.actor == null
+- Refactor regex pattern matching to utilities.js
+
 ## History
 
 0.9.4 - 5/06/2021